Extract expressions with all tests remaining working
